requirements.txt
fastapi
uvicorn[standard]
jinja2
sqlmodel
psycopg2-binary
python-dotenv
sentence-transformers
scikit-learn
numpy
pydantic
watchfiles
openaipydantic-settings